pyproject.toml
[tool.poetry]
name = "contact-tracker-2"
version = "0.1.0"
description = ""
authors = ["Joshua Huston <joshua.david.huston@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
